requirements.txt
anyio==4.11.0
asgiref==3.10.0
cachecontrol==0.14.3
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cloudinary==1.44.1
cryptography==46.0.3
dj-database-url==3.0.1
django==5.2.7
django-autoslug==1.9.9
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
django-phonenumber-field==8.3.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
firebase-admin==7.1.0
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.42.1
google-auth-httplib2==0.2.1
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
msgpack==1.1.2
packaging==25.0
phonenumbers==9.0.17
pillow==12.0.0
pip==25.3
proto-plus==1.26.1
protobuf==6.33.0
psycopg-binary==3.2.12
psycopg2==2.9.11
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pyjwt==2.10.1
pyparsing==3.2.5
pytz==2025.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing-extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
